+++ android/src/main/java/com/engagelab/privates/flutter_plugin_engagelab/MsgToJson.java
@@ -8,6 +8,7 @@ import com.engagelab.privates.push.api.InAppMessage;
 import com.engagelab.privates.push.api.NotificationMessage;
 import com.engagelab.privates.push.api.PlatformTokenMessage;
 import com.engagelab.privates.push.api.TagMessage;
+import com.engagelab.privates.push.api.VoipDataMessage;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -122,6 +123,19 @@ public class MsgToJson {
         return jsonObject.toString();
     }
 
+
+    public static String voipMessageToJson(VoipDataMessage voipDataMessage) {
+        JSONObject jsonObject = new JSONObject();
+        try {
+            jsonObject.put("messageId", voipDataMessage.getMessageId());
+            jsonObject.put("appkey", voipDataMessage.getAppkey());
+            jsonObject.put("extraData", voipDataMessage.getExtraData());
+            jsonObject.put("platform", voipDataMessage.getPlatform());
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+        return jsonObject.toString();
+    }
 
     public static String booleanToJson(boolean enable) {
         JSONObject jsonObject = new JSONObject();